Migliora gestione contenuti e permessi

+++ Backend/modules/authors/authors.controller.js
@@ -4,7 +4,7 @@ const authorsService = require("./authors.service")
 const getAuthors = async (req, res) => {
     const result = await authorsService.getAllAuthors(
         req.query.page,
-        req.query.limit
+        req.query.pageSize ?? req.query.limit
     )
 
     res.status(200).json({
@@ -14,16 +14,6 @@ const getAuthors = async (req, res) => {
         totalPages: result.totalPages,
         currentPage: result.currentPage,
         data: result.authors
-    })
-}
-
-const createAuthor = async (req, res) => {
-    const author = await authorsService.createAuthor(req.body)
-
-    res.status(201).json({
-        statusCode: 201,
-        message: "Created",
-        data: author
     })
 }
 
@@ -42,6 +32,10 @@ const getAuthor = async (req, res) => {
 }
 
 const updateAuthor = async (req, res) => {
+    if (req.author.id !== req.params.id) {
+        throw new AppError(403, "Non puoi modificare un altro autore")
+    }
+
     const author = await authorsService.updateAuthor(req.params.id, req.body)
 
     if (!author) {
@@ -56,6 +50,10 @@ const updateAuthor = async (req, res) => {
 }
 
 const deleteAuthor = async (req, res) => {
+    if (req.author.id !== req.params.id) {
+        throw new AppError(403, "Non puoi eliminare un altro autore")
+    }
+
     const author = await authorsService.deleteAuthor(req.params.id)
 
     if (!author) {
@@ -69,6 +67,10 @@ const deleteAuthor = async (req, res) => {
 }
 
 const uploadAvatar = async (req, res) => {
+    if (req.author.id !== req.params.authorId) {
+        throw new AppError(403, "Non puoi modificare un altro autore")
+    }
+
     if (!req.file) {
         throw new AppError(400, "Avatar file is required")
     }
@@ -89,11 +91,28 @@ const uploadAvatar = async (req, res) => {
     })
 }
 
+const updateAuthorRole = async (req, res) => {
+    const author = await authorsService.updateAuthorRole(
+        req.params.id,
+        req.body.role
+    )
+
+    if (!author) {
+        throw new AppError(404, "Author not found")
+    }
+
+    res.status(200).json({
+        statusCode: 200,
+        message: "Author role updated",
+        data: author
+    })
+}
+
 module.exports = {
     getAuthors,
-    createAuthor,
     getAuthor,
     updateAuthor,
     deleteAuthor,
-    uploadAvatar
+    uploadAvatar,
+    updateAuthorRole
 }
